move test_run_txs to logging_app.py
